Call `routerHeartbeats.Combined()` within Compute
One of our controllers crashed and printed the following:

```
fatal error: concurrent map iteration and map write
```

This was because routerHeartbeats.Combined() was looping over the
routerHeartbeats while handleHeartbeat was mutating the
routerHeartbeats. The fix is call routerHeartbeats.Combined() within
Compute so we have exclusive access to the routerHeartbeats when
Combined() is called.