Log more details about scaling decisions
To help us understand why functions are scaling up and down, and how much more skipper would use if it could, let's log more details about scaling decisions when they happen! This factors out a `ScalingDecision` struct describing all the inputs to the decision and the decision itself that we can then use to log.