Switch to using maxUnavailable for skipper pdbs
We want to disrupt skipper as little as possible when doing node scaleups/scaledowns to minimize the hash ring churn. Previously the skipper PDBs set minAvailable: <%= replica_count - 1 %> , which works great if we have a fixed replica count. But I think we've since added HPAs, and now the replica count floats up and down. I think we still want to be as minimally disruptive as possible, so instead we should say `maxUnavailable: 1` and only disrupt one pod at once ever, regardless of the current HPA replica count.